+++ CHANGELOG.md
@@ -13,7 +13,12 @@
 
 ### Added
 
-`LuidLowPart` and `LuidHighPart` parameters to the `Invoke-AdvApiLookupPrivilegeName` function.
+* `LuidLowPart` and `LuidHighPart` parameters to the `Invoke-AdvApiLookupPrivilegeName` function.
+* `Invoke-AdvApiOpenSCManager` function for calling the advapi32.dll library's `OpenSCManager` function.
+* `Invoke-AdvApiCloseServiceHandle` function for calling the advapi32.dll library's `CloseServiceHandle` function.
+* `Invoke-AdvApiOpenService` function for calling the advapi32.dll library's `OpenService` function.
+* `Invoke-AdvApiQueryServiceConfig` function for calling the advapi32.dll library's `QueryServiceconfig` function.
+* `Invoke-AdvApiQueryServiceConfig2` function for calling the advapi32.dll library's `QueryServiceconfig2` function.
 
 ### Changed